requirements.txt
pydantic==2.4
uvicorn==0.23.2
requests==2.31.0
supabase==1.2.0
fastapi==0.104.1
python-dotenv==1.0.0